+++ src/lib.rs
@@ -412,7 +412,6 @@ pub mod opaque;
 pub mod ciphersuite;
 mod envelope;
 mod hash;
-mod serialization;
 
 pub mod group;
 pub mod map_to_curve;
@@ -426,6 +425,8 @@ pub mod oprf;
 mod oprf;
 
 pub mod slow_hash;
+
+mod serialization;
 
 #[cfg(test)]
 mod tests;
